week1/rhnatyshyn/flask/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
pip~=25.1.1
Mako~=1.3.10
MarkupSafe~=3.0.2
typing_extensions~=4.15.0
click~=8.3.0
blinker~=1.9.0
Jinja2~=3.1.6
itsdangerous~=2.2.0
SQLAlchemy~=2.0.43
alembic~=1.16.5
greenlet~=3.2.4
psycopg2-binary~=2.9.10
Flask-Migrate~=4.1.0
python-decouple~=3.8